+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Features
+
+- Add `sentryslog` integration ([#865](https://github.com/getsentry/sentry-go/pull/865))
+
 - Always set Mechanism Type to generic ([#896](https://github.com/getsentry/sentry-go/pull/897))
 
 ### Misc